✏️ [Modify] FontSize enum을 Manager 폴더로 이동
