+++ Sodam/Sodam/Model/Manager/Font/FontManager.swift
@@ -1,0 +1,12 @@
+//
+//  FontManager.swift
+//  Sodam
+//
+//  Created by EMILY on 18/02/2025.
+//
+
+import Foundation
+
+final class FontManager {
+    
+}
