+++ Sodam/Sodam/Model/Manager/Font/FontSize.swift
@@ -1,0 +1,15 @@
+//
+//  FontSize.swift
+//  Sodam
+//
+//  Created by EMILY on 18/02/2025.
+//
+
+import Foundation
+
+enum FontSize {
+    static let title: CGFloat = 27
+    static let title2: CGFloat = 24
+    static let body: CGFloat = 16
+    static let timeStamp: CGFloat = 14
+}
